# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Galleton local API","version":"0.1.0","description":"Single-trust-domain, loopback-only session renewal. No browser CORS or remote listener. The API bearer token controls all sessions. No provider session lifetime guarantee."},"servers":[{"url":"http://127.0.0.1:8766"}],"security":[{"localBearer":[]}],"paths":{"/v1/health":{"get":{"operationId":"health","summary":"Authenticated process health","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"ok":{"type":"boolean"},"version":{"type":"string"}},"required":["ok","version"],"additionalProperties":false}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}}}},"/v1/providers":{"get":{"operationId":"providers","summary":"Provider names, kinds and allowlisted origins","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"providers":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"},"kind":{"type":"string","enum":["oauth2","http"]},"origins":{"type":"array","items":{"type":"string"}}},"required":["name","kind","origins"],"additionalProperties":false}}},"required":["providers"],"additionalProperties":false}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}}}},"/v1/sessions":{"get":{"operationId":"listSessions","summary":"List non-secret metadata","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"sessions":{"type":"array","items":{"$ref":"#/components/schemas/Metadata"}}},"required":["sessions"],"additionalProperties":false}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}}}},"/v1/sessions/{id}":{"get":{"operationId":"getSession","summary":"Read non-secret metadata","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metadata"}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}}},"put":{"operationId":"connect","summary":"Import once; replace=true explicitly reconnects","responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metadata"}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}}}},"delete":{"operationId":"forget","summary":"Delete local state, without provider revocation","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"deleted":{"type":"boolean"}},"required":["deleted"],"additionalProperties":false}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"}}]},"/v1/sessions/{id}/headers":{"post":{"operationId":"headers","summary":"Renew when due, then return origin-scoped authentication","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadersResult"}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadersRequest"}}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"}}]},"/v1/sessions/{id}/refresh":{"post":{"operationId":"refresh","summary":"Request renewal; terminal states and retry deadlines are respected","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metadata"}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"}}]},"/v1/sessions/{id}/cookies":{"post":{"operationId":"capture","summary":"Persist individual Set-Cookie values from an approved origin","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metadata"}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaptureRequest"}}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"}}]},"/v1/sessions/{id}/request":{"post":{"operationId":"managedRequest","summary":"Make one authenticated resource request, capture cookies, do not replay","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestResult"}}}},"default":{"description":"Sanitized error; see code and optional retry_at","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorEnvelope"}}}}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestInput"}}}}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"}}]}},"components":{"securitySchemes":{"localBearer":{"type":"http","scheme":"bearer","description":"Contents of the private api.token file; not a provider token."}},"schemas":{"Problem":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"retry_at":{"type":"string","format":"date-time"}},"required":["code","message"],"additionalProperties":false},"ErrorEnvelope":{"type":"object","properties":{"error":{"$ref":"#/components/schemas/Problem"}},"required":["error"],"additionalProperties":false},"Metadata":{"type":"object","properties":{"id":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"},"provider":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"},"status":{"type":"string"},"revision":{"type":"integer","minimum":1},"created_at":{"type":"string","format":"date-time"},"last_refresh":{"type":"string","format":"date-time"},"next_refresh":{"type":"string","format":"date-time"},"access_expires_at":{"type":"string","format":"date-time"},"cookie_count":{"type":"integer","minimum":0},"error":{"$ref":"#/components/schemas/Problem"}},"required":["id","provider","status","revision","created_at","cookie_count"],"additionalProperties":false},"Credentials":{"type":"object","properties":{"provider":{"type":"string","pattern":"^[A-Za-z0-9_-]{1,64}$"},"cookie_origin":{"type":"string","format":"uri"},"cookie_header":{"type":"string","maxLength":131072},"set_cookies":{"type":"array","items":{"type":"string"},"maxItems":256},"access_token":{"type":"string","maxLength":65536},"refresh_token":{"type":"string","maxLength":65536},"access_expires_at":{"type":"string","format":"date-time"},"secrets":{"type":"object","additionalProperties":{"type":"string"}},"replace":{"type":"boolean","default":false}},"required":["provider"],"additionalProperties":false},"HeadersRequest":{"type":"object","properties":{"url":{"type":"string","format":"uri"}},"required":["url"],"additionalProperties":false},"HeadersResult":{"type":"object","properties":{"headers":{"type":"object","additionalProperties":{"type":"string"}},"revision":{"type":"integer"},"expires_at":{"type":"string","format":"date-time"}},"required":["headers","revision"],"additionalProperties":false},"CaptureRequest":{"type":"object","properties":{"url":{"type":"string","format":"uri"},"set_cookie":{"type":"array","items":{"type":"string"},"maxItems":256}},"required":["url","set_cookie"],"additionalProperties":false},"RequestInput":{"type":"object","properties":{"url":{"type":"string","format":"uri"},"method":{"type":"string","enum":["GET","HEAD","POST","PUT","PATCH","DELETE","OPTIONS"],"default":"GET"},"headers":{"type":"object","additionalProperties":{"type":"string"}},"body_base64":{"type":"string","contentEncoding":"base64","description":"At most 1 MiB decoded."}},"required":["url"],"additionalProperties":false},"RequestResult":{"type":"object","properties":{"status":{"type":"integer","minimum":100,"maximum":599},"headers":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}},"body_base64":{"type":"string","contentEncoding":"base64","description":"At most 4 MiB decoded; not a streaming API."},"revision":{"type":"integer"}},"required":["status","headers","body_base64","revision"],"additionalProperties":false},"Empty":{"type":"object","properties":{},"required":[],"additionalProperties":false}}}}